+++ src/components/dialog/SaveAlertDialog.tsx
@@ -6,7 +6,7 @@ import {
   AlertDialogFooter,
   AlertDialogHeader,
   AlertDialogTitle,
-} from '@/components/ui/alert-dialog';
+} from '@/components/ui/alert-dialog.tsx';
 
 export default function SaveAlertDialog({ open, onOpenChange, onDiscard }) {
   return (
