+++ src/features/user/components/LoginModal.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState } from 'react';
 import {
   Dialog,
   DialogContent,
@@ -6,27 +6,26 @@ import {
   DialogHeader,
   DialogTitle,
   DialogTrigger,
-} from "@/components/ui/dialog";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { supabase } from "@/supabase.ts";
-import { useToast } from "@/hooks/use-toast.ts";
-import {AuthModalProps} from "@/components/modal/SignupModal.tsx";
-
+} from '@/components/ui/dialog.tsx';
+import { Button } from '@/components/ui/button.tsx';
+import { Input } from '@/components/ui/input.tsx';
+import { Label } from '@/components/ui/label.tsx';
+import { supabase } from '@/supabase.ts';
+import { useToast } from '@/hooks/use-toast.ts';
+import { AuthModalProps } from '@/features/user/components/SignupModal.tsx';
 
 const AUTH_ERROR_MESSAGES = {
-  TITLE: "로그인 실패",
-  SUPABASE_MESSAGE: "Invalid login credentials",
-  INVALID_CREDENTIALS: "비밀번호가 올바르지 않습니다. 다시 시도해주세요.",
-  UNKNOWN_ERROR: "로그인 중 오류가 발생했습니다. 다시 시도해주세요.",
+  TITLE: '로그인 실패',
+  SUPABASE_MESSAGE: 'Invalid login credentials',
+  INVALID_CREDENTIALS: '비밀번호가 올바르지 않습니다. 다시 시도해주세요.',
+  UNKNOWN_ERROR: '로그인 중 오류가 발생했습니다. 다시 시도해주세요.',
 };
 
-const LoginModal = ({ isOpen, onOpenChange } : AuthModalProps) => {
+const LoginModal = ({ isOpen, onOpenChange }: AuthModalProps) => {
   const { toast } = useToast();
 
-  const [email, setEmail] = useState<string>("");
-  const [password, setPassword] = useState<string>("");
+  const [email, setEmail] = useState<string>('');
+  const [password, setPassword] = useState<string>('');
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -40,7 +39,7 @@ const LoginModal = ({ isOpen, onOpenChange } : AuthModalProps) => {
       if (error) {
         if (error.message.includes(AUTH_ERROR_MESSAGES.SUPABASE_MESSAGE)) {
           toast({
-            variant: "destructive",
+            variant: 'destructive',
             title: AUTH_ERROR_MESSAGES.TITLE,
             description: AUTH_ERROR_MESSAGES.INVALID_CREDENTIALS,
           });
@@ -50,7 +49,7 @@ const LoginModal = ({ isOpen, onOpenChange } : AuthModalProps) => {
       onOpenChange(false);
     } catch (err) {
       toast({
-        variant: "destructive",
+        variant: 'destructive',
         title: AUTH_ERROR_MESSAGES.TITLE,
         description: AUTH_ERROR_MESSAGES.UNKNOWN_ERROR,
       });
@@ -88,11 +87,7 @@ const LoginModal = ({ isOpen, onOpenChange } : AuthModalProps) => {
             />
           </div>
           <div className="flex justify-end space-x-2">
-            <Button
-              type="button"
-              variant="outline"
-              onClick={() => onOpenChange(false)}
-            >
+            <Button type="button" variant="outline" onClick={() => onOpenChange(false)}>
               취소
             </Button>
             <Button type="submit">로그인</Button>
