+++ src/features/user/components/OverlappingAvatars.tsx
@@ -1,5 +1,5 @@
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar.tsx';
-import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
+import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip.tsx';
 import { User } from '@/types/project.model.tsx';
 
 export default function OverlappingAvatars({ users }) {
@@ -7,7 +7,7 @@ export default function OverlappingAvatars({ users }) {
     <div className="flex items-center">
       <div className="flex -space-x-3">
         <TooltipProvider>
-        {users.map((user, index) => (
+          {users.map((user, index) => (
             <Tooltip>
               <TooltipTrigger asChild>
                 <Avatar key={index} className="w-7 h-7 border-2 border-white shadow-md">
@@ -21,7 +21,7 @@ export default function OverlappingAvatars({ users }) {
                 <p>{user.name}</p>
               </TooltipContent>
             </Tooltip>
-        ))}
+          ))}
         </TooltipProvider>
       </div>
     </div>
