+++ src/features/user/components/UserMenu.tsx
@@ -1,16 +1,16 @@
-import React, { useState } from "react";
-import { User } from "lucide-react";
+import React, { useState } from 'react';
+import { User } from 'lucide-react';
 import {
   DropdownMenu,
   DropdownMenuTrigger,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuSeparator,
-} from "@/components/ui/dropdown-menu.tsx";
-import LoginModal from "@/components/modal/LoginModal.tsx";
-import SignupModal from "@/components/modal/SignupModal.tsx";
-import { useAuth } from "@/contexts/AuthContext.tsx";
-import { supabase } from "@/supabase.ts";
+} from '@/components/ui/dropdown-menu.tsx';
+import LoginModal from '@/features/user/components/LoginModal.tsx';
+import SignupModal from '@/features/user/components/SignupModal.tsx';
+import { useAuth } from '@/contexts/AuthContext.tsx';
+import { supabase } from '@/supabase.ts';
 
 const UserMenu = () => {
   const [showLoginModal, setShowLoginModal] = useState(false);
@@ -47,36 +47,21 @@ const UserMenu = () => {
               <DropdownMenuItem>프로필</DropdownMenuItem>
               <DropdownMenuItem>설정</DropdownMenuItem>
               <DropdownMenuSeparator />
-              <DropdownMenuItem
-                onClick={handleLogout}
-                disabled={isLoggingOut}
-                className="text-red-500"
-              >
-                {isLoggingOut ? "로그아웃 중..." : "로그아웃"}
+              <DropdownMenuItem onClick={handleLogout} disabled={isLoggingOut} className="text-red-500">
+                {isLoggingOut ? '로그아웃 중...' : '로그아웃'}
               </DropdownMenuItem>
             </>
           ) : (
             <>
-              <DropdownMenuItem onClick={() => setShowLoginModal(true)}>
-                로그인
-              </DropdownMenuItem>
-              <DropdownMenuItem onClick={() => setShowSignupModal(true)}>
-                회원가입
-              </DropdownMenuItem>
+              <DropdownMenuItem onClick={() => setShowLoginModal(true)}>로그인</DropdownMenuItem>
+              <DropdownMenuItem onClick={() => setShowSignupModal(true)}>회원가입</DropdownMenuItem>
             </>
           )}
         </DropdownMenuContent>
       </DropdownMenu>
 
-      {showLoginModal && (
-        <LoginModal isOpen={showLoginModal} onOpenChange={setShowLoginModal} />
-      )}
-      {showSignupModal && (
-        <SignupModal
-          isOpen={showSignupModal}
-          onOpenChange={setShowSignupModal}
-        />
-      )}
+      {showLoginModal && <LoginModal isOpen={showLoginModal} onOpenChange={setShowLoginModal} />}
+      {showSignupModal && <SignupModal isOpen={showSignupModal} onOpenChange={setShowSignupModal} />}
     </div>
   );
 };
