+++ src/pages/DetailPage.tsx
@@ -4,13 +4,13 @@ import { Progress } from '@/components/ui/progress';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { Users, FileText, GitBranch, Link, Calendar, ArrowUpRight } from 'lucide-react';
 import { useNavigate, useParams } from 'react-router-dom';
-import ProjectCard from '@/components/detail/ProgressCard.tsx';
-import ProgressCard from '@/components/detail/ProgressCard.tsx';
+import ProjectCard from '@/features/project/components/ProgressCard.tsx';
+import ProgressCard from '@/features/project/components/ProgressCard.tsx';
 import { motion } from 'framer-motion';
-import MemberCard from '@/components/detail/MemberCard.tsx';
-import TechStackCard from '@/components/detail/TechStackCard.tsx';
+import MemberCard from '@/features/project/components/MemberCard.tsx';
+import TechStackCard from '@/features/project/components/TechStackCard.tsx';
 import { useQuery } from '@tanstack/react-query';
-import { fetchProject, fetchProjects } from '@/fecthers/project/project.tsx';
+import { fetchProject, fetchProjects } from '@/features/project/fetchers/project.tsx';
 import { useEffect, useState } from 'react';
 import { ProjectDTO } from '@/fecthers/project/project.dto.ts';
 import { Project } from '@/types/project.model.tsx';
